+++ packages/client/components/DashNavList/LeftNavItem.tsx
@@ -5,5 +5,13 @@ interface Props {
 }
 export const LeftNavItem = (props: Props) => {
   const {children} = props
-  return <div className='flex flex-1 flex-col font-medium text-sm'>{children}</div>
+  return (
+    <div
+      className={
+        'flex flex-1 flex-col overflow-hidden text-clip whitespace-nowrap font-medium text-sm'
+      }
+    >
+      {children}
+    </div>
+  )
 }